// src/app/models/con-voucher.item.ts
export class ConVoucherItem {
    numVoucher?: string;
    numItem?: string;
    codCuenta?: string;
    desCuenta?: string;
    codAuxiliar?: string;
    desAuxiliar?: string;
    numRuc?: string;
    codMoneda?: string;
    desMoneda?: string;
    codDocumento?: string;
    numSerieDoc?: string;
    numDocumentos?: string;
    indDebeHaber?: string;
    impSoles?: number;
    impDolares?: number;
}